requirements.txt
# requirements.txt - Consistent versions for all services

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Security & Authentication
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
cryptography==41.0.7
pyjwt==2.8.0

# Data Validation (Pydantic v2)
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Redis (for sessions/caching)
redis==5.0.1
hiredis==2.3.2

# Email
aiosmtplib==3.0.1
jinja2==3.1.2

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Environment Management
python-dotenv==1.0.0

# Utilities
pytz==2023.3
python-dateutil==2.8.2
typing-extensions==4.8.0

# Rate Limiting
slowapi==0.1.9

# Logging
loguru==0.7.2

# Development Tools (optional, for development environment)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Production Server
gunicorn==21.2.0

# Security
itsdangerous==2.1.2

# Database Utilities
greenlet==3.0.1

# Type Checking (optional)
types-redis==4.6.0.11
types-passlib==1.7.7.13
types-python-jose==3.3.4.8